Keep signed-out Svelte routes behind an explicit sign-in fallback
