+++ starters/sveltekit-betterauth/src/lib/JazzClientProvider.svelte
@@ -3,11 +3,12 @@
   import { JazzProvider, betterAuth } from "jazz-tools/svelte";
   import { env } from "$env/dynamic/public";
   import { authClient } from "$lib/auth-client";
+  import SignIn from "$lib/SignIn.svelte";
 
   let { children: pageChildren }: { children?: Snippet } = $props();
 </script>
 
 <JazzProvider appId={env.PUBLIC_JAZZ_APP_ID!} serverUrl={env.PUBLIC_JAZZ_SERVER_URL!} auth={betterAuth(authClient)}>
-  {#snippet signedOut()}{@render pageChildren?.()}{/snippet}
+  {#snippet signedOut()}<SignIn />{/snippet}
   {#snippet children()}{@render pageChildren?.()}{/snippet}
 </JazzProvider>
